skills/github-webhooks/examples/fastapi/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
python-dotenv>=1.0.0
